requirements.txt
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.2
django-adaptors==0.2.5
django-cors-headers==3.0.2
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2==2.8.3
py==1.8.0
pytz==2019.1
retry==0.9.2
sqlparse==0.3.0
whitenoise==4.1.2
